+++ sdk/php/tests/Structures/VehicleModelTest.php
@@ -17,14 +17,16 @@ class VehicleModelTest extends AbstractStructureTestCase
     public function testConfigure()
     {
         $this->instance = $this->factory($input = [
-            'id'      => $description = 'some description',
-            'name'    => $name = 'some name',
-            'mark_id' => $mark_id = 'some mark id',
+            'id'           => $description = 'some description',
+            'name'         => $name = 'some name',
+            'mark_id'      => $mark_id = 'some mark id',
+            'vehicle_type' => $model_vehicle_type = 'some model type id',
         ]);
 
         $this->assertEquals($description, $this->instance->getId());
         $this->assertEquals($name, $this->instance->getName());
         $this->assertEquals($mark_id, $this->instance->getMarkId());
+        $this->assertEquals($model_vehicle_type, $this->instance->getVehicleType());
 
         $this->assertEquals($input, $this->instance->toArray());
     }
@@ -35,14 +37,16 @@ class VehicleModelTest extends AbstractStructureTestCase
     public function testConfigureWithNulls()
     {
         $this->instance = $this->factory($input = [
-            'id'      => null,
-            'name'    => null,
-            'mark_id' => null,
+            'id'           => null,
+            'name'         => null,
+            'mark_id'      => null,
+            'vehicle_type' => null,
         ]);
 
         $this->assertNull($this->instance->getId());
         $this->assertNull($this->instance->getName());
         $this->assertNull($this->instance->getMarkId());
+        $this->assertNull($this->instance->getVehicleType());
     }
 
     /**